Crate js_parser

Javascript parser with aim to fully support Ecma script 2019 and Estree API. One of the core values of project is to represent ECMAScript standart as close as possible, which should improve redability and maintainability of the code.

This project is still a work in progress.

Lexer carries out transformation of input to linear stream of tokens. Rules for lexer closely resemble Lexical grammar in ESCMAScript. Lexer transform n letters of input to exactly 1 token. This allows better speed optimization due to no dynamic memory allocation.

To convert string to stream of tokens we can use code simillar to this

use javascript_lexer::Lexer;

static JS: &str = r#"
class A extend B {
    get a() {}
}
"#;

let tokens = Lexer::lex_tokens(JS).unwrap();

Parser constructs Abstract syntax tree (AST) from token stream and then checks AST for violation of static semantics rules. Namely Early errors which have to be reported before evaluation.

Construct AST from token stream

use javascript_lexer::Lexer;
use js_parser::{estree, Parser};

static JS: &str = r#"class A extend B {get a() {}}"#;

let tokens = Lexer::lex_tokens(JS).unwrap();
Parser::ast_tree(tokens, estree::ProgramSourceType::Script);

Re-exports

pub extern crate javascript_lexer;

Modules

estree

node
This file contains definition of AST structures. Structures are grouped to enums by their semantic value. Most important enums are: Expression, Statement, Node, Literal, Declaration. Same node can be contained in multiple enums.

static_semantics

Structs

Parser